Online Fitness App/fitness_app/requirements.txt
Django==4.2.20
djangorestframework==3.14.0